+++ default/role.defender.js
@@ -2,7 +2,7 @@ module.exports = {
 
   run: creep => {
 
-    if (creep.suicideAt(40)) return;
+    if (creep.recycleAt(40)) return;
 
     if (!creep.isCharged()) {